Move home related classes to the .home package
