+++ android_mobile/src/main/java/com/alexstyl/specialdates/upcoming/widget/list/WidgetRouterActivity.kt
@@ -11,8 +11,8 @@ import com.alexstyl.specialdates.date.Date
 import com.alexstyl.specialdates.date.getDateExtraOrThrow
 import com.alexstyl.specialdates.date.putExtraDate
 import com.alexstyl.specialdates.events.namedays.activity.NamedayActivity
+import com.alexstyl.specialdates.home.HomeActivity
 import com.alexstyl.specialdates.person.PersonActivity
-import com.alexstyl.specialdates.upcoming.UpcomingEventsActivity
 import javax.inject.Inject
 
 private val ACTION_VIEW_NAMEDAY = "action:view_nameday"
@@ -62,7 +62,7 @@ class WidgetRouterActivity : Activity() {
         ACTION_VIEW_CONTACT -> PersonActivity.buildIntentFor(context, intent.contact())
         ACTION_VIEW_NAMEDAY -> NamedayActivity.getStartIntent(context, intent.getDateExtraOrThrow())
         else -> {
-            UpcomingEventsActivity.getStartIntent(this)
+            HomeActivity.getStartIntent(this)
         }
     }
